Completed day 5 part 1, stuck in part 2
Part 2 works in theory but practically takes up too much memory because the number
ranges for the seeds is too large - leading to too much memory being created.
The method needs to be updated to only keep track of ranges as we go through the maps.
This would mean the ranges are possibly divided into smaller and smaller ranges as they
are processed by more and more maps. The processing of each map also has to be updated to
now work with ranges.